refactoring and made Edit component
