+++ clientSide/src/components/IndividualBooking/IndividualBooking.css
@@ -35,3 +35,7 @@ tbody tr:nth-child(even) td {
   cursor: pointer;
   color: var(--global-green-63);
 }
+.fa-edit {
+  cursor: pointer;
+  color: var(--global-green-63);
+}
